requirements.txt
chardet==2.3.0
jinja2==2.8n